+++ components/bot_detector/core/structs/kafka/scraped.py
@@ -1,7 +1,8 @@
-from .._metadata import MetaData
-from bot_detector.highscore_worker.structs import HighscoreBaseStruct
+from bot_detector.highscore.structs import HighscoreBaseStruct
 from bot_detector.player.structs import PlayerStruct
 from pydantic import BaseModel
+
+from .._metadata import MetaData
 
 
 class ScrapedStruct(BaseModel):
