--- components/bot_detector/highscore_worker/database/__init__.py
@@ -1,17 +1,0 @@
-"""Highscore worker persistence helpers."""
-
-from .repository import HighscoreDataRepo
-from .structs import (
-    HighscoreDataDailyTableStruct,
-    HighscoreDataLatestTableStruct,
-    HighscoreDataMonthlyTableStruct,
-    HighscoreDataWeeklyTableStruct,
-)
-
-__all__ = [
-    "HighscoreDataRepo",
-    "HighscoreDataDailyTableStruct",
-    "HighscoreDataWeeklyTableStruct",
-    "HighscoreDataMonthlyTableStruct",
-    "HighscoreDataLatestTableStruct",
-]
